Obtain microbe signatures from BugSigDB

Functionality for obtaining microbe signatures from BugSigDB

getSignatures(
  df,
  tax.id.type = c("ncbi", "metaphlan", "taxname"),
  tax.level = "mixed",
  exact.tax.level = TRUE,
  min.size = 1
)

Arguments

df
data.frame storing BugSigDB data. Typically obtained via importBugSigDB.

tax.id.type
Character. Taxonomic ID type of the returned microbe sets. Either "ncbi" (default), "metaphlan", or "taxname".

tax.level
character. Either "mixed" or any subset of c("kingdom", "phylum", "class", "order", "family", "genus", "species", "strain"). This full vector is equivalent to "mixed".

exact.tax.level
logical. Should only the exact taxonomic level specified by tax.level be returned? Defaults to TRUE. If FALSE, a more general tax.level is extracted for microbes given at a more specific taxonomic level.

min.size
integer. Minimum signature size. Defaults to 1, which will filter out empty signature. Use min.size = 0 to keep empty signatures.

Value

a list of microbe signatures. Each signature is a character vector of taxonomic IDs depending on the chosen tax.id.type.

References

BugSigDB: https://bugsigdb.org

See also

Examples

 df <- importBugSigDB()
#> Using cached version from 2021-09-01 12:47:31
 sigs <- getSignatures(df)