Ontology-based subsetting of BugSigDB signatures

This function facilitates ontology-based queries for experimental factors and body sites.

subsetByOntology(df, column = c("Body site", "Condition"), term, ontology)

Arguments

df

data.frame storing BugSigDB data. Typically obtained via importBugSigDB.

column

character. Column of df on which subsetting should be performed. Should be either "Body site" (default) or "Condition".

term

character. A valid term of ontology. Subsetting by this term then involves subsetting column to all descendants of that term in the the chosen ontology and that are present in the chosen column of df.

ontology

an object of class ontology_index as defined in the ontologyIndex package. Typically obtained via getOntology.

Value

a data.frame with the chosen column restricted to descendants of the chosen term in the chosen ontology.

References

EFO: https://www.ebi.ac.uk/ols/ontologies/efo

UBERON: https://www.ebi.ac.uk/ols/ontologies/uberon

See also

Examples

 # (1) Obtain BugSigDB data
 df <- importBugSigDB()
#> Using cached version from 2023-02-20 10:51:00

 # (2) Obtain ontology of interest as an R object
 uberon <- getOntology("uberon")
#> Using cached version from 2023-02-20 10:51:42

 # (3) High-level query on body site
 sdf <- subsetByOntology(df, 
                         column = "Body site",
                         term = "digestive system element",
                         ontology = uberon)
 table(sdf[,"Body site"])
#> 
#>                          Major salivary gland 
#>                                             1 
#>                                        caecum 
#>                                            15 
#>                                         colon 
#>                                            19 
#>                                colonic mucosa 
#>                                             2 
#>               feces,mucosa of small intestine 
#>                                             4 
#>    feces,stomach,caecum,small intestine,colon 
#>                                             2 
#>                                         ileum 
#>                                             2 
#> mouth,skin of cheek,tongue,gingiva,oropharynx 
#>                                             2 
#>               mucosa of small intestine,feces 
#>                                             3 
#>                             mucosa of stomach 
#>                                             3 
#>                                    oral gland 
#>                                            51 
#>                                        rectum 
#>                                            24 
#>                        saliva-secreting gland 
#>                                             2 
#>                               small intestine 
#>                                            10 
#>                                       stomach 
#>                                            41 
#>                                        tongue 
#>                                            28