Obtain meta-signatures for a column of interest

Functionality for obtaining meta-signatures for a column of interest

getMetaSignatures(
  df,
  column,
  direction = c("BOTH", "UP", "DOWN"),
  min.studies = 2,
  min.taxa = 5,
  comb.fun = sum,
  ...
)

Arguments

df

data.frame storing BugSigDB data. Typically obtained via importBugSigDB.

column

character. Column of interest. Need to be a valid column name of df.

direction

character. Indicates direction of abundance change for signatures to be included in the computation of meta-signatures. Use "UP" to restrict computation to signatures with increased abundance in the exposed group. Use "DOWN" to restrict to signatures with decreased abundance in the exposed group. Defaults to "BOTH" which will not filter signatures by direction of abundance change.

min.studies

integer. Minimum number of studies for a category in column to be included. Defaults to 2, which will then only compute meta-signatures for categories investigated by at least two studies.

min.taxa

integer. Minimum size for meta-signatures. Defaults to 5, which will then only include meta-signatures containing at least 5 taxa.

comb.fun

function. Function for combining sample size of the exposed group and sample size of the unexposed group into an overall study sample size. Defaults to sum which will simply add sample sizes of exposed and unexposed group.

...

additionals argument passed on to getSignatures.

Value

A list of meta-signatures, each meta-signature being a named numeric vector. Names are the taxa of the meta-signature, numeric values correspond to sample size weights associated with each taxon.

See also

Examples

 df <- importBugSigDB()
#> Using cached version from 2024-09-16 11:05:46

 # Body-site specific meta-signatures composed from signatures reported as both 
 # increased or decreased across all conditions of study:
 bs.meta.sigs <- getMetaSignatures(df, column = "Body site")

 # Condition-specific meta-signatures from fecal samples, increased
 # in conditions of study. Use taxonomic names instead of the default NCBI IDs:
 df.feces <- df[df$`Body site` == "Feces", ]
 cond.meta.sigs <- getMetaSignatures(df.feces, column = "Condition", 
                                     direction = "UP", tax.id.type = "taxname")

 # Inspect the results
 names(cond.meta.sigs)
#>  [1] "Acute lymphoblastic leukemia"                             
#>  [2] "Acute myeloid leukemia"                                   
#>  [3] "Age"                                                      
#>  [4] "Air pollution"                                            
#>  [5] "Alzheimer's disease"                                      
#>  [6] "Anorexia nervosa"                                         
#>  [7] "Antimicrobial agent"                                      
#>  [8] "Anxiety disorder"                                         
#>  [9] "Asthma"                                                   
#> [10] "Atopic eczema"                                            
#> [11] "Attention deficit hyperactivity disorder"                 
#> [12] "Autism"                                                   
#> [13] "Bipolar disorder"                                         
#> [14] "Breast cancer"                                            
#> [15] "COVID-19"                                                 
#> [16] "Cervical cancer"                                          
#> [17] "Cesarean section"                                         
#> [18] "Chronic fatigue syndrome"                                 
#> [19] "Chronic kidney disease"                                   
#> [20] "Colorectal adenoma"                                       
#> [21] "Colorectal cancer"                                        
#> [22] "Crohn's disease"                                          
#> [23] "Diabetes mellitus"                                        
#> [24] "Diarrhea"                                                 
#> [25] "Diet"                                                     
#> [26] "Eczema"                                                   
#> [27] "Endometriosis"                                            
#> [28] "Epilepsy"                                                 
#> [29] "Ethnic group"                                             
#> [30] "Exercise"                                                 
#> [31] "Food allergy"                                             
#> [32] "Gastric cancer"                                           
#> [33] "Gestational diabetes"                                     
#> [34] "Graft versus host disease"                                
#> [35] "Gut microbiome measurement"                               
#> [36] "HIV infection"                                            
#> [37] "Hashimoto's thyroiditis"                                  
#> [38] "Hepatocellular carcinoma"                                 
#> [39] "High fat diet"                                            
#> [40] "Irritable bowel syndrome"                                 
#> [41] "Leukemia"                                                 
#> [42] "Lung cancer"                                              
#> [43] "Milk allergic reaction"                                   
#> [44] "Multiple myeloma"                                         
#> [45] "Multiple sclerosis"                                       
#> [46] "Non-Hodgkins lymphoma"                                    
#> [47] "Non-alcoholic fatty liver disease"                        
#> [48] "Obesity"                                                  
#> [49] "Pancreatic carcinoma"                                     
#> [50] "Parkinson's disease"                                      
#> [51] "Physical activity"                                        
#> [52] "Polycystic ovary syndrome"                                
#> [53] "Psoriasis"                                                
#> [54] "Response to allogeneic hematopoietic stem cell transplant"
#> [55] "Response to immunochemotherapy"                           
#> [56] "Rheumatoid arthritis"                                     
#> [57] "Schizophrenia"                                            
#> [58] "Socioeconomic status"                                     
#> [59] "Treatment"                                                
#> [60] "Type I diabetes mellitus"                                 
#> [61] "Type II diabetes mellitus"                                
#> [62] "Ulcerative colitis"                                       
#> [63] "Unipolar depression"                                      
#> [64] "Urinary tract infection"                                  
 cond.meta.sigs["Bipolar disorder"]
#> $`Bipolar disorder`
#>    Flavonifractor   Lachnospiraceae    Actinomycetota Coriobacteriaceae 
#>         0.3459215         0.1767372         0.1193353         0.1193353 
#>  Coriobacteriales    Coriobacteriia 
#>         0.1193353         0.1193353 
#>