Obtain meta-signatures for a column of interest

Functionality for obtaining meta-signatures for a column of interest

getMetaSignatures(
  df,
  column,
  direction = c("BOTH", "UP", "DOWN"),
  min.studies = 2,
  min.taxa = 5,
  comb.fun = sum,
  ...
)

Arguments

df

data.frame storing BugSigDB data. Typically obtained via importBugSigDB.

column

character. Column of interest. Need to be a valid column name of df.

direction

character. Indicates direction of abundance change for signatures to be included in the computation of meta-signatures. Use "UP" to restrict computation to signatures with increased abundance in the exposed group. Use "DOWN" to restrict to signatures with decreased abundance in the exposed group. Defaults to "BOTH" which will not filter signatures by direction of abundance change.

min.studies

integer. Minimum number of studies for a category in column to be included. Defaults to 2, which will then only compute meta-signatures for categories investigated by at least two studies.

min.taxa

integer. Minimum size for meta-signatures. Defaults to 5, which will then only include meta-signatures containing at least 5 taxa.

comb.fun

function. Function for combining sample size of the exposed group and sample size of the unexposed group into an overall study sample size. Defaults to sum which will simply add sample sizes of exposed and unexposed group.

...

additionals argument passed on to getSignatures.

Value

A list of meta-signatures, each meta-signature being a named numeric vector. Names are the taxa of the meta-signature, numeric values correspond to sample size weights associated with each taxon.

See also

Examples

 df <- importBugSigDB()
#> Using cached version from 2022-10-29 05:45:22
 # Body-site specific meta-signatures composed from signatures reported as both 
 # increased or decreased across all conditions of study:
 bs.meta.sigs <- getMetaSignatures(df, column = "Body site")
 # Condition-specific meta-signatures from feces samples, increased
 # in conditions of study. Use taxonomic names instead of the default NCBI IDs:
 df.feces <- df[df$`Body site` == "feces", ]
 cond.meta.sigs <- getMetaSignatures(df.feces, column = "Condition", 
                     direction = "UP", tax.id.type = "taxname")
 names(cond.meta.sigs)
#>  [1] "COVID-19"                                                 
#>  [2] "Eczema"                                                   
#>  [3] "HIV infection"                                            
#>  [4] "Hashimoto's thyroiditis"                                  
#>  [5] "Parkinson's disease"                                      
#>  [6] "acute lymphoblastic leukemia"                             
#>  [7] "age"                                                      
#>  [8] "air pollution"                                            
#>  [9] "anorexia nervosa"                                         
#> [10] "antimicrobial agent"                                      
#> [11] "anxiety disorder"                                         
#> [12] "asthma"                                                   
#> [13] "atopic eczema"                                            
#> [14] "attention deficit hyperactivity disorder"                 
#> [15] "autism"                                                   
#> [16] "bipolar disorder"                                         
#> [17] "cervical cancer"                                          
#> [18] "cesarean section"                                         
#> [19] "chronic fatigue syndrome"                                 
#> [20] "chronic kidney disease"                                   
#> [21] "colorectal cancer"                                        
#> [22] "diet"                                                     
#> [23] "endometriosis"                                            
#> [24] "epilepsy"                                                 
#> [25] "food allergy"                                             
#> [26] "gastric cancer"                                           
#> [27] "gestational diabetes"                                     
#> [28] "graft versus host disease"                                
#> [29] "irritable bowel syndrome"                                 
#> [30] "leukemia"                                                 
#> [31] "lung cancer"                                              
#> [32] "milk allergic reaction"                                   
#> [33] "multiple myeloma"                                         
#> [34] "multiple sclerosis"                                       
#> [35] "non-Hodgkins lymphoma"                                    
#> [36] "non-alcoholic fatty liver disease"                        
#> [37] "obesity"                                                  
#> [38] "pancreatic carcinoma"                                     
#> [39] "response to allogeneic hematopoietic stem cell transplant"
#> [40] "rheumatoid arthritis"                                     
#> [41] "schizophrenia"                                            
#> [42] "type I diabetes mellitus"                                 
#> [43] "type II diabetes mellitus"                                
#> [44] "unipolar depression"                                      
#> [45] "urinary tract infection"                                  
 cond.meta.sigs["bipolar disorder"]
#> $`bipolar disorder`
#>    Flavonifractor   Lachnospiraceae    Actinobacteria Coriobacteriaceae 
#>         0.3459215         0.1767372         0.1193353         0.1193353 
#>  Coriobacteriales    Coriobacteriia 
#>         0.1193353         0.1193353 
#>